approx: OCCT `int` degree fields modelled signed (data-model alignment) - fixes the nbp-5 underflow in Approx_ComputeLine
Symptom: several draft_angle / feat cases panicked with "attempt to
subtract with overflow" at approx_int.rs (m_degmax = nbp - 5).

OCCT source: Approx_ComputeLine.gxx L1336 `int nbp = lpt - fpt + 1;` and
L1344-1350
  int Mdegmax = mydegremax;
  if (nbp < Mdegmax + 5 && mycut) { Mdegmax = nbp - 5; }
  if (Mdegmax < mydegremin)       { Mdegmax = mydegremin; }
so `nbp - 5` is legitimately NEGATIVE for short point ranges and the very
next statement clamps it back to mydegremin - the negative value is part of
the design and never escapes. A usize model underflowed instead, and could
not represent the intermediate.

Aligned data model (OCCT declarations read first):
- ComputeLine::{mydegremin,mydegremax} -> i32 (Approx_ComputeLine.hxx `int`),
  new/init parameters too;
- GeomInt_WLApprox::{myDegMin,myDegMax} -> i32 (GeomInt_WLApprox.hxx L160-161
  are `int`), set_parameters parameters too;
- BRepApprox_Approx::SetParameters deg_min/deg_max -> i32
  (BRepApprox_Approx.hxx L103-111 are `const int`), with the ds_filler.rs
  dmin/dmax locals;
- IntTools_FaceFace's approx_parameters_for (cxx L2736-2783 ApproxParameters)
  -> i32;
- both `let nbp = lpt - fpt + 1;` -> `lpt as i32 - fpt as i32 + 1`;
- at the rcad APIs that index with usize (Gradient::new, MultiCurve::new,
  set_value, make_ml_between) the OCCT int is cast explicitly with a note
  (the loop guarantees deg >= 0 and nbp >= 1).

Measured: draft_angle's library panic count drops 28 -> 25 (two approx_int
sites and brep_offset_api_draft_angle.rs:83 leave the panic chain); pass
counts unchanged (0/49) because the remaining cases fail at state asserts or
the two draft_modification EMap-geometry walls.

Gates: rcad-algo 415/0/0, rcad-kernel 689/0, tktopalgo 36/36, pavefiller
26/26, builder_stage 76/76 + smoke 1/1; eight grids 8/8 after rebuilding the
exes; every domain grid unchanged in pass counts. Probes: 0.

Lesson (same family as pitfall 22): porting an OCCT `int` field as Rust
`usize` is not just style - it loses the signed intermediate semantics.
Read the OCCT declaration before choosing the type.

+++ libs/rcad-algo/src/bop/int_tools/face_make_curve.rs
@@ -1556,10 +1556,10 @@ fn make_part_curve(
 
 /// OCCT IntTools_FaceFace::ApproxParameters (IntTools_FaceFace.cxx
 /// L2736-2783): degree and iteration parameters of the WLine approximation.
-fn approx_parameters_for(surf1: &Surface3, surf2: &Surface3) -> (usize, usize, i32) {
+fn approx_parameters_for(surf1: &Surface3, surf2: &Surface3) -> (i32, i32, i32) {
     let mut i_nb_iter = 0i32;
-    let mut i_deg_min = 4usize;
-    let mut i_deg_max = 8usize;
+    let mut i_deg_min = 4i32;
+    let mut i_deg_max = 8i32;
     // Cylinder/Torus.
     let cyl_rad = match (surf1, surf2) {
         (Surface3::Cylinder(c), Surface3::Torus(t)) => Some((c.radius, t.minor_radius)),
